refactor: cleanup `CommonJsExportsParserPlugin`
