+++ crates/rspack_plugin_javascript/src/parser_plugin/common_js_exports_parse_plugin.rs
@@ -1,0 +1,532 @@
+use rspack_core::{
+  extract_member_expression_chain, BuildMetaDefaultObject, BuildMetaExportsType, RuntimeGlobals,
+  RuntimeRequirementsDependency, SpanExt,
+};
+use swc_core::atoms::Atom;
+use swc_core::common::Spanned;
+use swc_core::ecma::ast::{AssignExpr, CallExpr, PropOrSpread};
+use swc_core::ecma::ast::{Callee, ExprOrSpread, Ident, MemberExpr, ObjectLit};
+use swc_core::ecma::ast::{Expr, Lit, Pat, PatOrExpr, Prop, PropName, ThisExpr, UnaryOp};
+
+use super::JavascriptParserPlugin;
+use crate::dependency::{CommonJsExportRequireDependency, CommonJsExportsDependency};
+use crate::dependency::{CommonJsSelfReferenceDependency, ExportsBase, ModuleDecoratorDependency};
+use crate::visitors::{expr_matcher, JavascriptParser};
+
+const MODULE_NAME: &str = "module";
+const EXPORTS_NAME: &str = "exports";
+
+fn get_member_expression_info(
+  expr: &Expr,
+  is_module_exports_start: Option<bool>,
+) -> Option<Vec<Atom>> {
+  let is_module_exports_start = match is_module_exports_start {
+    Some(v) => v,
+    None => is_module_exports_member_expr_start(expr),
+  };
+  expr.as_member().and_then(|expr: &MemberExpr| {
+    let members = extract_member_expression_chain(expr)
+      .members()
+      .iter()
+      .skip(if is_module_exports_start { 2 } else { 1 })
+      .map(|n| n.0.to_owned())
+      .collect::<Vec<_>>();
+    match expr.obj {
+      box Expr::Call(_) => Some(members),
+      box Expr::Ident(_) => Some(members),
+      box Expr::MetaProp(_) => Some(members),
+      box Expr::This(_) => Some(members),
+      _ if expr_matcher::is_module_exports(&expr.obj) => Some(members),
+      _ => None,
+    }
+  })
+}
+
+fn is_module_exports_member_expr_start(mut expr: &Expr) -> bool {
+  loop {
+    match expr {
+      _ if expr_matcher::is_module_exports(expr) => return true,
+      Expr::Member(MemberExpr { obj, .. }) => expr = obj.as_ref(),
+      _ => return false,
+    }
+  }
+}
+
+fn get_value_of_property_description(expr_or_spread: &ExprOrSpread) -> Option<&Expr> {
+  if let ExprOrSpread {
+    expr: box Expr::Object(ObjectLit { props, .. }),
+    ..
+  } = expr_or_spread
+  {
+    for prop in props {
+      if let PropOrSpread::Prop(prop) = prop
+        && let Prop::KeyValue(key_value_prop) = &**prop
+        && let PropName::Ident(ident) = &key_value_prop.key
+        && &ident.sym == "value"
+      {
+        return Some(&key_value_prop.value);
+      }
+    }
+  }
+  None
+}
+
+fn is_truthy_literal(expr: &Expr) -> bool {
+  match expr {
+    Expr::Lit(lit) => is_lit_truthy_literal(lit),
+    Expr::Unary(unary) => {
+      if unary.op == UnaryOp::Bang {
+        return is_falsy_literal(&unary.arg);
+      }
+      false
+    }
+    _ => false,
+  }
+}
+
+fn is_falsy_literal(expr: &Expr) -> bool {
+  match expr {
+    Expr::Lit(lit) => !is_lit_truthy_literal(lit),
+    Expr::Unary(unary) => {
+      if unary.op == UnaryOp::Bang {
+        return is_truthy_literal(&unary.arg);
+      }
+      false
+    }
+    _ => false,
+  }
+}
+
+fn is_lit_truthy_literal(lit: &Lit) -> bool {
+  match lit {
+    Lit::Str(str) => !str.value.is_empty(),
+    Lit::Bool(bool) => bool.value,
+    Lit::Null(_) => false,
+    Lit::Num(num) => num.value != 0.0,
+    _ => true,
+  }
+}
+
+impl<'parser> JavascriptParser<'parser> {
+  fn is_exports_member_expr_start(&mut self, mut expr: &Expr) -> bool {
+    loop {
+      match expr {
+        _ if self.is_exports_expr(expr) => return true,
+        Expr::Member(MemberExpr { obj, .. }) => expr = obj.as_ref(),
+        _ => return false,
+      }
+    }
+  }
+
+  fn is_module_ident(&mut self, ident: &Ident) -> bool {
+    ident.sym == MODULE_NAME && self.is_unresolved_ident(MODULE_NAME)
+  }
+
+  fn is_exports_ident(&mut self, ident: &Ident) -> bool {
+    ident.sym == EXPORTS_NAME && self.is_unresolved_ident(EXPORTS_NAME)
+  }
+
+  fn is_exports_expr(&mut self, expr: &Expr) -> bool {
+    matches!(expr,  Expr::Ident(ident) if self.is_exports_ident(ident))
+  }
+
+  fn is_top_level_this(&self, _expr: &ThisExpr) -> bool {
+    self.top_level_scope
+  }
+
+  fn is_top_level_this_expr(&self, expr: &Expr) -> bool {
+    matches!(expr,  Expr::This(e) if self.is_top_level_this(e))
+  }
+
+  fn is_exports_or_module_exports_or_this_expr(&mut self, expr: &Expr) -> bool {
+    self.is_exports_expr(expr)
+      || expr_matcher::is_module_exports(expr)
+      || self.is_top_level_this_expr(expr)
+  }
+
+  // can't scan `__esModule` value
+  fn bailout(&mut self) {
+    if matches!(self.parser_exports_state, Some(true)) {
+      self.build_meta.exports_type = BuildMetaExportsType::Unset;
+      self.build_meta.default_object = BuildMetaDefaultObject::False;
+    }
+    *self.parser_exports_state = Some(false);
+  }
+
+  // `__esModule` is false
+  fn enable(&mut self) {
+    if matches!(self.parser_exports_state, Some(false)) {
+      return;
+    }
+    if self.parser_exports_state.is_none() {
+      self.build_meta.exports_type = BuildMetaExportsType::Default;
+      self.build_meta.default_object = BuildMetaDefaultObject::Redirect;
+    }
+    *self.parser_exports_state = Some(true);
+  }
+
+  // `__esModule` is true
+  fn set_flagged(&mut self) {
+    if matches!(self.parser_exports_state, Some(false)) || self.parser_exports_state.is_none() {
+      return;
+    }
+    if matches!(self.build_meta.exports_type, BuildMetaExportsType::Dynamic) {
+      return;
+    }
+    self.build_meta.exports_type = BuildMetaExportsType::Flagged;
+  }
+
+  // `__esModule` is dynamic, eg `true && true`
+  fn set_dynamic(&mut self) {
+    if matches!(self.parser_exports_state, Some(false)) || self.parser_exports_state.is_none() {
+      return;
+    }
+    self.build_meta.exports_type = BuildMetaExportsType::Dynamic;
+  }
+
+  fn check_namespace(&mut self, top_level: bool, value_expr: Option<&Expr>) {
+    if matches!(self.parser_exports_state, Some(false)) || self.parser_exports_state.is_none() {
+      return;
+    }
+    if let Some(value_expr) = value_expr
+      && is_truthy_literal(value_expr)
+      && top_level
+    {
+      self.set_flagged();
+    } else {
+      self.set_dynamic();
+    }
+  }
+
+  fn is_this_member_expr_start(&self, mut expr: &Expr) -> bool {
+    if self.enter_call != 0 {
+      return false;
+    }
+    loop {
+      match expr {
+        _ if self.is_top_level_this_expr(expr) => return true,
+        Expr::Member(MemberExpr { obj, .. }) => expr = obj.as_ref(),
+        _ => return false,
+      }
+    }
+  }
+
+  fn is_require_call(&mut self, node: &CallExpr) -> bool {
+    node
+      .callee
+      .as_expr()
+      .map(|expr| matches!(expr, box Expr::Ident(ident) if &ident.sym == "require" && self.is_unresolved_ident("require")))
+      .unwrap_or_default()
+  }
+
+  fn is_require_call_expr(&mut self, expr: &Expr) -> bool {
+    matches!(expr, Expr::Call(call_expr) if self.is_require_call(call_expr))
+  }
+
+  // FIXME: this function should be deleted because it just a hack
+  fn append_module_runtime(&mut self) {
+    self
+      .presentational_dependencies
+      .push(Box::new(RuntimeRequirementsDependency::new(
+        RuntimeGlobals::MODULE,
+      )));
+  }
+}
+
+pub struct CommonJsExportsParserPlugin;
+
+impl JavascriptParserPlugin for CommonJsExportsParserPlugin {
+  fn identifier(&self, parser: &mut JavascriptParser, ident: &Ident) -> Option<bool> {
+    if parser.is_module_ident(ident) {
+      parser.append_module_runtime();
+      // here should use, but scanner is not one pass, so here use extra `visit_program` to calculate is_harmony
+      // matches!( self.build_meta.exports_type, BuildMetaExportsType::Namespace)
+      let decorator = if parser.is_esm {
+        RuntimeGlobals::HARMONY_MODULE_DECORATOR
+      } else {
+        RuntimeGlobals::NODE_MODULE_DECORATOR
+      };
+      parser.bailout();
+      parser
+        .presentational_dependencies
+        .push(Box::new(ModuleDecoratorDependency::new(decorator)));
+      Some(true)
+    } else if !parser.is_esm && parser.is_exports_ident(ident) {
+      parser.bailout();
+      parser
+        .dependencies
+        .push(Box::new(CommonJsSelfReferenceDependency::new(
+          (ident.span().real_lo(), ident.span().real_hi()),
+          ExportsBase::Exports,
+          vec![],
+          false,
+        )));
+      Some(true)
+    } else {
+      None
+    }
+  }
+
+  fn this(
+    &self,
+    parser: &mut JavascriptParser,
+    expr: &swc_core::ecma::ast::ThisExpr,
+  ) -> Option<bool> {
+    if parser.is_esm {
+      None
+    } else if parser.is_top_level_this(expr) {
+      parser.bailout();
+      parser
+        .dependencies
+        .push(Box::new(CommonJsSelfReferenceDependency::new(
+          (expr.span().real_lo(), expr.span().real_hi()),
+          ExportsBase::This,
+          vec![],
+          false,
+        )));
+      Some(true)
+    } else {
+      None
+    }
+  }
+
+  fn member(&self, parser: &mut JavascriptParser, mem_expr: &MemberExpr) -> Option<bool> {
+    if parser.is_esm {
+      return None;
+    }
+
+    // FIXME: remove this `.clone`
+    let expr = Expr::Member(mem_expr.clone());
+    let handle_remaining = |parser: &mut JavascriptParser, base: ExportsBase| {
+      let is_module_exports_start = matches!(base, ExportsBase::ModuleExports);
+      if let Some(remaining) = get_member_expression_info(&expr, Some(is_module_exports_start)) {
+        if remaining.is_empty() {
+          parser.bailout();
+        }
+        parser
+          .dependencies
+          .push(Box::new(CommonJsSelfReferenceDependency::new(
+            (expr.span().real_lo(), expr.span().real_hi()),
+            base,
+            remaining,
+            false,
+          )));
+        Some(true)
+      } else {
+        None
+      }
+    };
+    if expr_matcher::is_module_exports(&expr) {
+      // `module.exports`
+      parser.bailout();
+      parser
+        .dependencies
+        .push(Box::new(CommonJsSelfReferenceDependency::new(
+          (expr.span().real_lo(), expr.span().real_hi()),
+          ExportsBase::ModuleExports,
+          vec![],
+          false,
+        )));
+      Some(true)
+    } else if parser.is_exports_member_expr_start(&expr) {
+      // `exports.x.y`
+      handle_remaining(parser, ExportsBase::Exports)
+    } else if is_module_exports_member_expr_start(&expr) {
+      // `module.exports.x.y`
+      parser.append_module_runtime();
+      handle_remaining(parser, ExportsBase::ModuleExports)
+    } else if parser.is_this_member_expr_start(&expr) {
+      // `this.x.y`
+      handle_remaining(parser, ExportsBase::This)
+    } else {
+      None
+    }
+  }
+
+  fn assign(&self, parser: &mut JavascriptParser, assign_expr: &AssignExpr) -> Option<bool> {
+    if parser.is_esm {
+      return None;
+    }
+    let PatOrExpr::Pat(box Pat::Expr(box left_expr)) = &assign_expr.left else {
+      return None;
+    };
+
+    let handle_remaining = |parser: &mut JavascriptParser, base: ExportsBase| {
+      let is_module_exports_start = matches!(base, ExportsBase::ModuleExports);
+      if let Some(remaining) = get_member_expression_info(left_expr, Some(is_module_exports_start))
+      {
+        parser.walk_expression(&assign_expr.right);
+        if remaining.is_empty() {
+          parser.enable();
+
+          if parser.is_require_call_expr(&assign_expr.right) {
+            // exports = require('xx');
+            // module.exports = require('xx');
+            // this = require('xx');
+            // It's possible to reexport __esModule, so we must convert to a dynamic module
+            parser.set_dynamic();
+            let related_require_dep = parser
+              .dependencies
+              .iter()
+              .find(|item| item.is_span_equal(&assign_expr.right.span()))
+              .map(|item| item.id())
+              .cloned();
+            parser
+              .dependencies
+              .push(Box::new(CommonJsExportRequireDependency::new(
+                (left_expr.span().real_lo(), left_expr.span().real_hi()),
+                base,
+                remaining,
+                related_require_dep,
+              )));
+          } else {
+            // exports = {};
+            // module.exports = {};
+            // this = {};
+            parser.bailout();
+          }
+        } else {
+          // exports.__esModule = true;
+          // module.exports.__esModule = true;
+          // this.__esModule = true;
+          if let Some(first_member) = remaining.first()
+            && first_member == "__esModule"
+          {
+            parser.check_namespace(
+              // const flagIt = () => (exports.__esModule = true); => stmt_level = 1, last_stmt_is_expr_stmt = false
+              // const flagIt = () => { exports.__esModule = true }; => stmt_level = 2, last_stmt_is_expr_stmt = true
+              // (exports.__esModule = true); => stmt_level = 1, last_stmt_is_expr_stmt = true
+              parser.stmt_level == 1 && parser.last_stmt_is_expr_stmt,
+              Some(&assign_expr.right),
+            );
+          }
+          // exports.a = 1;
+          // module.exports.a = 1;
+          // this.a = 1;
+          parser
+            .dependencies
+            .push(Box::new(CommonJsExportsDependency::new(
+              (left_expr.span().real_lo(), left_expr.span().real_hi()),
+              None,
+              base,
+              remaining.to_owned(),
+            )));
+        }
+
+        Some(true)
+      } else {
+        None
+      }
+    };
+
+    if parser.is_exports_member_expr_start(left_expr) {
+      // exports.x = y;
+      parser.enable();
+      handle_remaining(parser, ExportsBase::Exports)
+    } else if is_module_exports_member_expr_start(left_expr) {
+      // module.exports.x = y;
+      parser.append_module_runtime();
+      handle_remaining(parser, ExportsBase::ModuleExports)
+    } else if parser.is_this_member_expr_start(left_expr) {
+      // this.x = y
+      handle_remaining(parser, ExportsBase::This)
+    } else {
+      None
+    }
+  }
+
+  fn call(&self, parser: &mut JavascriptParser, call_expr: &CallExpr) -> Option<bool> {
+    if parser.is_esm {
+      None
+    } else if let Callee::Expr(expr) = &call_expr.callee {
+      let handle_remaining = |parser: &mut JavascriptParser, base: ExportsBase| {
+        let is_module_exports_start = matches!(base, ExportsBase::ModuleExports);
+        if let Some(remaining) = get_member_expression_info(expr, Some(is_module_exports_start)) {
+          // exports()
+          // module.exports()
+          // this()
+          if remaining.is_empty() {
+            parser.bailout();
+          }
+
+          // exports.a.b()
+          // module.exports.a.b()
+          // this.a.b()
+          parser
+            .dependencies
+            .push(Box::new(CommonJsSelfReferenceDependency::new(
+              (expr.span().real_lo(), expr.span().real_hi()),
+              base,
+              remaining,
+              true,
+            )));
+          parser.walk_expr_or_spread(&call_expr.args);
+          Some(true)
+        } else {
+          None
+        }
+      };
+      // Object.defineProperty(exports, "xxx", { value: 1 });
+      // Object.defineProperty(module.exports, "xxx", { value: 1 });
+      // Object.defineProperty(this, "xxx", { value: 1 });
+      if expr_matcher::is_object_define_property(expr)
+        && let Some(ExprOrSpread { expr, .. }) = call_expr.args.first()
+        && parser.is_exports_or_module_exports_or_this_expr(expr)
+        && let Some(arg2) = call_expr.args.get(2)
+      {
+        parser.enable();
+
+        if let Some(ExprOrSpread {
+          expr: box Expr::Lit(Lit::Str(str)),
+          ..
+        }) = call_expr.args.get(1)
+        {
+          // Object.defineProperty(exports, "__esModule", { value: true });
+          // Object.defineProperty(module.exports, "__esModule", { value: true });
+          // Object.defineProperty(this, "__esModule", { value: true });
+          if str.value == "__esModule" {
+            parser.check_namespace(
+              parser.stmt_level == 1,
+              get_value_of_property_description(arg2),
+            );
+          }
+
+          let base = if parser.is_exports_expr(expr) {
+            ExportsBase::DefinePropertyExports
+          } else if expr_matcher::is_module_exports(expr) {
+            ExportsBase::DefinePropertyModuleExports
+          } else if parser.is_top_level_this_expr(expr) {
+            ExportsBase::DefinePropertyThis
+          } else {
+            panic!("Unexpected expr type");
+          };
+          parser
+            .dependencies
+            .push(Box::new(CommonJsExportsDependency::new(
+              (call_expr.span.real_lo(), call_expr.span.real_hi()),
+              Some((arg2.span().real_lo(), arg2.span().real_hi())),
+              base,
+              vec![str.value.clone()],
+            )));
+        }
+
+        parser.walk_expression(&arg2.expr);
+        Some(true)
+      } else if parser.is_exports_member_expr_start(expr) {
+        // exports.x()
+        handle_remaining(parser, ExportsBase::Exports)
+      } else if is_module_exports_member_expr_start(expr) {
+        // module.exports.x()
+        parser.append_module_runtime();
+        handle_remaining(parser, ExportsBase::ModuleExports)
+      } else if parser.is_this_member_expr_start(expr) {
+        // this.x()
+        handle_remaining(parser, ExportsBase::This)
+      } else {
+        None
+      }
+    } else {
+      None
+    }
+  }
+}
